Initial Coding
Create the coding to complete the assignment- Opens, reads, and loops through the lines of the file. Skips any lines that do no begin with ‘X-DSPAM-Confidence:’. When the line begins with ‘X-DSPAM-Confidence:’, slices to the data we need and strips whitespace. Counts the number of iterations and sums the total by the new float number found in each line. Returns an average based on the total divided by the count.